Make git lab fork work for projects you are a member of
The git lab fork command will fail to fork a project which you are a member of
as it finds that you have a project with the same name, even though it belongs
to a different user.

Alter the method used to check that do not already have a fork of the project.