GitSpindle.repository(): use the repo root, not the cwd, to determine the repo name
When a remote has not been configured for a spindle, a last resort guess
is done using the rpo directory name. But this mistakenly used the name
of the current working directory, not the repository root.

This is hard to test for in create.t, but will be tested in the next
patch, which exercises cat/ls in a subdir for repos that are missing
remotes, forcing the guess above to be done.