issue #118: Make links in rendered GitHub documents work
The rendered document that is returned by GitHub has all anchor names
prefixed with "user-content-". The links to those anchors are unchanged
though. If the page is displayed on GitHub, there exists some magic
JavaScript that handles the clicks on those links and translates them
accordingly so that the respective target is jumped to.

As the documents rendered with git-spindle are only meant to provide a
preview to see and test the result before checking in the Markdown file,
now all "user-content-" prefixes are stripped before displaying the
rendered page, so that all the links work natively in the result file.