issue #116: change temp file handling for GitHub render to make it work platform independently
As per the API documentation of tempfile.NamedTemporaryFile, it is
guaranteed that the generated file is present on the filesystem with the
name you can query. But it is not defined nor guaranteed what happens if
you try to open the generated file while the file handle is still open.

On Windows the problem is, that the file is locked and the browser cannot
open it as long as the temporary file handle is not closed and as soon as
the file handle is closed, the file gets deleted. At this point the
browser could read the file, but usually it vanished too fast for the
browser to read and render the temporary file.

Now the temporary file will not be deleted anymore automatically after the
file descriptor is closed, the file descriptor is closed before the
browser is called to render the file and after the one second grace period
sleep to let the browser render the file, it is removed manually.
This way the displaying of the rendered file should work platform
independently.