add 'git <spindle> help <command>' to display the help for only one subcommand
If you want to know the parameters and options of a specific git-spindle
sub-command, it is very cumbersome to display the full help and search for
the right sub-command.

Now there is a dedicated "git <spindle> help <command>" sub-command for
each spindle that displays only the help part about the actual command
requested.