issue #119: set remote.pushDefault in a triangular setup and make the user able to specify an upstream branch to track, also fall back to track fork branch if upstream branch was not found
Up to now only existing branches were affected by the --triangular
parameter. Now also remote.pushDefault is set to origin, so that also
future branches are by default pushed in the triangular way to origin.

Furthermore up to now if --triangular is in effect a tracking branch was
only set when a matching branch in upstream was found. Now you can specify
with --upstream-branch which branch in upstream should be tracked instead
of a name-matching one. If no branch in upstream is found that should be
tracked, as fallback a name-matching branch in origin is searched and
tracked if found. --upstream-branch is for example needed, if you have
several separate branches for different pull requests and want all to
track upstream/master which is their target branch.