{pull,merge}-request: merge into the branch we're tracking
If the branch you are creating a pull-request for tracks some upstream
branch, it is most likely that branch you want your changes pulled into.
For example you create a local branch from upstream/master, do some
changes and want them to be pulled, you want them in upstream/master. If
you create a local branch from upstream/v5.0.x and do some changes, it is
most likely the v5.0.x branch you want your changes to be pulled in.

Now the tracked branch in upstream is the default branch a pr is created
for and not anymore the default branch of the upstream repository which
might even be totally incompatible with your changes as it was not the
branch the changes were developed and tested against. If there is no
branch tracked or the tracked branch is not in the pull-target, the
default still is the default branch of the pull-target which most often is
"master" usually.