issue #121: fix exception during backing up a message due to an exception, e. g. during pull request submitting
When backing up a user-supplied message - e. g. the pull request message -
when an exception occured, a byte array was supplied to the write method,
but the write method expects a str. Now the "msg" str is given to the
method directly. Otherwise another exception is thrown and the message is
not saved properly.