fix the handling of GitHub issue and pull request templates and use the pull-request template from parent repo, not from the fork
find_template returned a byte array, and so when a template was present,
the string concatenating threw a "TypeError: can't concat bytes to str".
As the gh._session.get return value already has the decoded message as str
in its text property, this is used now instead of trying to decode
manually.

The pull request template files were not taken from the repository you
want to pull into, but from your fork. This somehow contradicts the
purpose of those files, especially if you do not maintain the master
branch in your fork, but only the dedicated pull request branches. Now the
template is taken from the repository you want to pull into.

When a pull request template was found and used, the title property was
not set and raised an exception when trying to use it later on. Now the
title property gets properly set in all cases.