unset the repo parameter if it was recognized as issue number
If the repo parameter is recognized as numeric value, it is assumed the
repo parameter is left out and it is an issue number instead. The previous
code added the recognized number to the list of issues, but did not set
the repo parameter to None. This way `git <spindle> issue 1` searched for
the issue number 1 in the repository <user>/1 instead of in the current
repository. If a repo actually consists only of digits, it must be
qualified with the username to make it non-numerical and thus recognized
as repository rather than an issue id.