fix ANSI color output
On some terminals it happened that colored output was not properly readable, because background color and foreground color were the same when using some attributes.
This change resets the background and foreground colors to default before applying the custom colors or attributes which makes the colored output being displayed properly.