fix hook editing
When editing a hook you got a complaint that the argument to
'properties/config' is not an object and that the argument to
'properties/events' is not an array. The name as first parameter was
plainly wrong and thus the parameters were shifted by one and their types
not like expected.