always set the fetch refspec for GitHub pull requests
Up to now the refspec was only set when issues were enabled and the repository was not a fork.
Both restrictions do not make much sense.
If issues are not enabled, there will simply be nothing to fetch, but as soon as they are enabled,
they will be available on fetch without the need to set-origin again.
And if the repo is a fork, nothing prevents others to open pull requests between different forks,
you are not restricted to creating pull-requests for the original repository.