make issues without repo but with filter or query work
According to the usage string, you should be able to specify a filter (for hub and lab) or query (for bb) without specifying a repo on calling git spindle issues.
This case is now detected by the presence of an equals sign inside the repo.
If one is found, the repos is seen as first filter, or as query string if no other query string is given.