improve the login procedures
Up to now the login procedures did not handle some cases nicely, like a revoked token, an empty username or password, a credential helper that also asks the user for the username, a credential helper that delivers a token with too few scopes and so on. Now the login procedures take care nicely of all sorts of edge cases and also prevents the usage of the Git Credential Manager for Windows. The latter is necessary unfortunately. The nice thing with the Git Credential Manager for Windows is, that it stores the credentials safely in the Windows Credentials Store, but the bad thing is, that it has special treatment for GitHub where it automatically requests and delivers a token with too few scopes for git-spindle and in newer versions also for BitBucket where it requests a token with too few scopes if two-factor authorization is enabled for the user.