fix the calendar commands
For GitHub the calculation whether the first month label should be displayed or not was erroneous. Furthermore depending on the encoding of stdout, the command did not work at all, because the character that is used for first of month was not displayable. Now there is a safety check that only displayable characters are used, as fallback "x" is shown. If the characters for first of month and rest are the same, the first of month now gets a red background to distinguish it better.

For GitLab the same problems that were present for GitHub were there. Additionally the calendar calculation did not work at all anymore as GitLab seems to have changed how they represent the data. According to the code they had an entry for each day of the last year just like GitHub has, but now only days that have commits are included in the data.