fix forking of GitLab repositories
According to GitLab support forking on GitLab is a two-step process. Forking via API copies the metadata, but the actual code fork is done asynchronously. This second step can need some time, depending on size of the repo and load of the systems. It is also a requirement to be able to fetch or clone. Unfortunately there is no way to watch the fork status according to GitLab support, so try for 2 minutes once a second to clone or fetch the new fork. Until then the code fork should usually be finished.