fix repo test-cleanup for GitHub
GitHub has a bug that the /user/repos endpoint of the API does need some seconds or even minutes to recognize that a repo was deleted. Until then it still lists the repo while it is listed nowhere else. Work around this bug while GitHub is working on a fix by waiting for 2 minutes and checking once a second whether the /user/repos endpoint is clean regarding the deleted repos.