wait for deletion to complete for create and fork
If you try to create or fork a repository, while the old repository at those coordinates is still being deleted on GitLab due to its asynchronous working mode, the operation fails with an according error message from GitLab.
It is now tried for 120 times, waiting a second between the tries to repeat the operation if it fails due to that exact reason.
An end-user would see the recommendation to try again later and could do so manually, but this way it is more convenient.
Where this change is important and essential, are the tests. Since a recent update to GitLab, the tests almost never run through due to this behavior. With this change they are working again.