Repo selection: give 'origin' a higher precedence
When trying to determine which remote to use for git hub et al., the
order used to be "mine" -> "first repo found". This is now "mine" ->
"origin (if on the same platform)" -> "first repo found".

This helps people who originally cloned a fork, renamed the remote and
added the upstream as "origin" later.